Move utils folder to places folder It was only being imported in the places handlers file, so I thought to make it immediately accessible this way.
