+++ routes/places/utils/createCity.js
@@ -1,8 +1,8 @@
 const axios = require('axios');
 
-const keys = require('../../config/keys');
+const keys = require('../../../config/keys');
 
-const City = require('../../models/City');
+const City = require('../../../models/City');
 
 module.exports = async (city, state, region, country) => {
   // Get city coordinates
